refac(subscriber): move common code out of layer module
The `layer` module included a lot of code which would be reused by a
second layer which would write a chunked file. It's possible that not
everything can be reused, but for now it will all be moved into a
`common` module.
