+++ rfr-subscriber/src/subscriber/mod.rs
@@ -1,3 +1,4 @@
+mod common;
 mod layer;
 
 pub use layer::RfrLayer;
